+++ java-service-examples/src/test/java/nbbrd/service/examples/FileTypeSpiTest.java
@@ -1,7 +1,7 @@
 package nbbrd.service.examples;
 
 import internal.FileTypeSpiLoader;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.atIndex;
@@ -10,10 +10,8 @@ public class FileTypeSpiTest {
 
     @Test
     public void test() {
-        FileTypeSpiLoader x = new FileTypeSpiLoader();
-
-        assertThat(x.get())
-                .containsExactlyElementsOf(x.get())
+        assertThat(FileTypeSpiLoader.get())
+                .containsExactlyElementsOf(FileTypeSpiLoader.get())
                 .hasSize(2)
                 .satisfies(o -> assertThat(o).isInstanceOf(FileTypeSpi.ByMagicNumberProbe.class), atIndex(0))
                 .satisfies(o -> assertThat(o).isInstanceOf(FileTypeSpi.ByExtensionProbe.class), atIndex(1));