+++ java-service-processor/src/main/java/internal/nbbrd/service/HasMethod.java
@@ -1,0 +1,14 @@
+package internal.nbbrd.service;
+
+import org.checkerframework.checker.nullness.qual.NonNull;
+
+import javax.lang.model.element.ExecutableElement;
+
+public interface HasMethod {
+
+    @NonNull ExecutableElement getMethod();
+
+    default @NonNull String getMethodName() {
+        return getMethod().getSimpleName().toString();
+    }
+}