+++ java-service-processor/src/main/java/internal/nbbrd/service/Wrapper.java
@@ -1,33 +1,34 @@
 /*
  * Copyright 2019 National Bank of Belgium
- * 
- * Licensed under the EUPL, Version 1.1 or - as soon they will be approved 
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
- * 
+ *
  * http://ec.europa.eu/idabc/eupl
- * 
- * Unless required by applicable law or agreed to in writing, software 
+ *
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package internal.nbbrd.service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import org.checkerframework.checker.nullness.qual.NonNull;
+
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.Modifier;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.util.ElementFilter;
 import javax.lang.model.util.Types;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
- *
  * @author Philippe Charles
  */
 @lombok.Value
@@ -43,7 +44,8 @@ public class Wrapper {
                         .filter(method -> isOneArgPublicMethod(util, method, service))
                         .map(Element.class::cast);
             }
-        }, STATIC_METHOD {
+        },
+        STATIC_METHOD {
             @Override
             public Stream<Element> parse(Types util, TypeElement service, TypeElement provider) {
                 return ElementFilter
@@ -56,13 +58,16 @@ public class Wrapper {
             }
         };
 
-        public abstract Stream<Element> parse(Types util, TypeElement service, TypeElement provider);
+        public abstract @NonNull Stream<Element> parse(@NonNull Types util, @NonNull TypeElement service, @NonNull TypeElement provider);
     }
 
+    @lombok.NonNull
     Kind kind;
+
+    @lombok.NonNull
     Element element;
 
-    public static List<Wrapper> allOf(Types util, TypeElement service, TypeElement provider) {
+    public static @NonNull List<Wrapper> allOf(@NonNull Types util, @NonNull TypeElement service, @NonNull TypeElement provider) {
         return Stream.of(Kind.values())
                 .flatMap(kind -> kind.parse(util, service, provider).map(element -> new Wrapper(kind, element)))
                 .collect(Collectors.toList());