+++ java-service-processor/src/main/java/internal/nbbrd/service/definition/Lifecycle.java
@@ -1,33 +1,33 @@
 /*
  * Copyright 2019 National Bank of Belgium
- * 
- * Licensed under the EUPL, Version 1.1 or - as soon they will be approved 
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
- * 
+ *
  * http://ec.europa.eu/idabc/eupl
- * 
- * Unless required by applicable law or agreed to in writing, software 
+ *
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package internal.nbbrd.service.definition;
 
 import internal.nbbrd.service.Unreachable;
 import nbbrd.service.Mutability;
+import org.checkerframework.checker.nullness.qual.NonNull;
 
 /**
- *
  * @author Philippe Charles
  */
 enum Lifecycle {
 
     IMMUTABLE, MUTABLE, CONCURRENT, CONSTANT, ATOMIC, UNSAFE_MUTABLE;
 
-    static Lifecycle of(Mutability mutability, boolean singleton) {
+    static @NonNull Lifecycle of(@NonNull Mutability mutability, boolean singleton) {
         switch (mutability) {
             case NONE:
                 return singleton ? CONSTANT : IMMUTABLE;
@@ -85,7 +85,7 @@ enum Lifecycle {
         }
     }
 
-    Mutability toMutability() {
+    @NonNull Mutability toMutability() {
         switch (this) {
             case IMMUTABLE:
             case CONSTANT: