+++ java-service-processor/src/main/java/internal/nbbrd/service/definition/LoadSorter.java
@@ -1,39 +1,49 @@
 /*
  * Copyright 2019 National Bank of Belgium
- * 
- * Licensed under the EUPL, Version 1.1 or - as soon they will be approved 
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
- * 
+ *
  * http://ec.europa.eu/idabc/eupl
- * 
- * Unless required by applicable law or agreed to in writing, software 
+ *
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package internal.nbbrd.service.definition;
 
-import java.util.Optional;
+import internal.nbbrd.service.HasMethod;
+import org.checkerframework.checker.index.qual.NonNegative;
+
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.TypeElement;
+import java.util.Optional;
 
 /**
- *
  * @author Philippe Charles
  */
 @lombok.Value
-class LoadSorter {
+class LoadSorter implements HasMethod {
 
     public enum KeyType {
         DOUBLE, INT, LONG, COMPARABLE
     }
 
+    @lombok.NonNull
     ExecutableElement method;
+
     boolean reverse;
+
+    @NonNegative
     int position;
+
+    @lombok.NonNull
     Optional<KeyType> keyType;
+
+    @lombok.NonNull
     Optional<TypeElement> serviceType;
 }