+++ java-service-processor/src/main/java/internal/nbbrd/service/provider/ServiceProviderProcessor.java
@@ -25,6 +25,7 @@ import javax.annotation.processing.SupportedAnnotationTypes;
 import javax.lang.model.SourceVersion;
 import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.Modifier;
+import javax.lang.model.element.Name;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
@@ -41,6 +42,10 @@ import java.util.stream.Stream;
 @SupportedAnnotationTypes({"nbbrd.service.ServiceProvider", "nbbrd.service.ServiceProvider.List"})
 public final class ServiceProviderProcessor extends AbstractProcessor {
 
+    // we store Name instead of TypeElement due to a bug(?) in JDK8
+    // that creates not-equivalent elements between rounds
+    private final Set<Name> pendingRefs = new HashSet<>();
+
     @Override
     public SourceVersion getSupportedSourceVersion() {
         return SourceVersion.latestSupported();
@@ -48,20 +53,47 @@ public final class ServiceProviderProcessor extends AbstractProcessor {
 
     @Override
     public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment roundEnv) {
+        pushPending(new AnnotationRegistry(annotations, roundEnv).readAll());
+
         if (roundEnv.processingOver()) {
-            return false;
+            List<ProviderRef> annotationRefs = popPending();
+
+            try {
+                checkRefs(annotationRefs);
+                writeRefs(annotationRefs);
+            } catch (RefError ex) {
+                reportError(ex.ref, ex.getMessage());
+            }
         }
 
-        List<ProviderRef> annotationRefs = new AnnotationRegistry(annotations, roundEnv).readAll();
+        return false;
+    }
 
-        if (annotationRefs.isEmpty()) {
-            return true;
-        }
+    private void pushPending(List<ProviderRef> refs) {
+        refs.forEach(ref -> pendingRefs.add(ref.getProvider().getQualifiedName()));
+    }
 
+    private List<ProviderRef> popPending() {
+        List<ProviderRef> result = pendingRefs
+                .stream()
+                .map(x -> processingEnv.getElementUtils().getTypeElement(x.toString()))
+                .flatMap(AnnotationRegistry::newRefs)
+                .collect(Collectors.toList());
+        pendingRefs.clear();
+        return result;
+    }
+
+    private void checkRefs(List<ProviderRef> refs) {
+        ProviderRef.getDuplicates(refs)
+                .forEach(ref -> {
+                    throw new RefError(ref, String.format("Duplicated provider: '%1$s'", ref.getProvider()));
+                });
+
+        refs.forEach(this::checkRef);
+    }
+
+    private void writeRefs(List<ProviderRef> annotationRefs) {
         log("Annotation", annotationRefs);
-
-        checkRefs(annotationRefs);
-
         try {
             checkModulePath(annotationRefs, new ModulePathRegistry(processingEnv));
             registerClassPath(annotationRefs, new ClassPathRegistry(processingEnv));
@@ -69,15 +101,6 @@ public final class ServiceProviderProcessor extends AbstractProcessor {
             String msg = ex.getClass().getSimpleName() + ": " + ex.getMessage();
             processingEnv.getMessager().printMessage(Diagnostic.Kind.ERROR, msg);
         }
-
-        return true;
-    }
-
-    private void checkRefs(List<ProviderRef> refs) {
-        ProviderRef.getDuplicates(refs)
-                .forEach(ref -> error(ref, String.format("Duplicated provider: '%1$s'", ref.getProvider())));
-
-        refs.forEach(this::checkRef);
     }
 
     private void checkRef(ProviderRef ref) {
@@ -85,28 +108,23 @@ public final class ServiceProviderProcessor extends AbstractProcessor {
         Types types = processingEnv.getTypeUtils();
 
         if (types.isSameType(ref.getService().asType(), elements.getTypeElement(Void.class.getName()).asType())) {
-            error(ref, "Cannot infer service from provider " + ref.getProvider());
-            return;
+            throw new RefError(ref, "Cannot infer service from provider " + ref.getProvider());
         }
 
         if (!types.isAssignable(ref.getProvider().asType(), types.erasure(ref.getService().asType()))) {
-            error(ref, String.format("Provider '%1$s' doesn't extend nor implement service '%2$s'", ref.getProvider(), ref.getService()));
-            return;
+            throw new RefError(ref, String.format("Provider '%1$s' doesn't extend nor implement service '%2$s'", ref.getProvider(), ref.getService()));
         }
 
         if (ref.getProvider().getEnclosingElement().getKind() == ElementKind.CLASS && !ref.getProvider().getModifiers().contains(Modifier.STATIC)) {
-            error(ref, String.format("Provider '%1$s' must be static inner class", ref.getProvider()));
-            return;
+            throw new RefError(ref, String.format("Provider '%1$s' must be static inner class", ref.getProvider()));
         }
 
         if (ref.getProvider().getModifiers().contains(Modifier.ABSTRACT)) {
-            error(ref, String.format("Provider '%1$s' must not be abstract", ref.getProvider()));
-            return;
+            throw new RefError(ref, String.format("Provider '%1$s' must not be abstract", ref.getProvider()));
         }
 
         if (Instantiator.allOf(types, ref.getService(), ref.getProvider()).stream().noneMatch(this::isValidInstantiator)) {
-            error(ref, String.format("Provider '%1$s' must have a public no-argument constructor", ref.getProvider()));
-            return;
+            throw new RefError(ref, String.format("Provider '%1$s' must have a public no-argument constructor", ref.getProvider()));
         }
     }
 
@@ -121,6 +139,10 @@ public final class ServiceProviderProcessor extends AbstractProcessor {
         }
     }
 
+    private void reportError(ProviderRef ref, String message) {
+        processingEnv.getMessager().printMessage(Diagnostic.Kind.ERROR, message, ref.getProvider());
+    }
+
     private void log(String id, List<ProviderRef> refs) {
         refs.forEach(ref -> processingEnv.getMessager().printMessage(Diagnostic.Kind.NOTE, String.format("%1$s: %2$s", id, ref)));
     }
@@ -129,8 +151,14 @@ public final class ServiceProviderProcessor extends AbstractProcessor {
         entries.forEach(entry -> processingEnv.getMessager().printMessage(Diagnostic.Kind.NOTE, String.format("%1$s: %2$s", id, entry)));
     }
 
-    private void error(ProviderRef ref, String message) {
-        processingEnv.getMessager().printMessage(Diagnostic.Kind.ERROR, message, ref.getProvider());
+    static class RefError extends IllegalArgumentException {
+
+        final ProviderRef ref;
+
+        RefError(ProviderRef ref, String message) {
+            super(message);
+            this.ref = ref;
+        }
     }
 
     private void errorEntry(ProviderEntry entry, String message) {
@@ -147,10 +175,12 @@ public final class ServiceProviderProcessor extends AbstractProcessor {
             logEntries("ModulePath", modulePathEntries);
 
             getMissingRefs(annotationRefs, modulePathEntries)
-                    .forEachOrdered(ref -> error(ref, "Missing module-info.java 'provides' directive for '" + ref + "'"));
+                    .forEachOrdered(ref -> {
+                        throw new RefError(ref, "Missing module-info.java 'provides' directive for '" + ref + "'");
+                    });
 
-            System.out.println(Arrays.toString(annotationRefs.toArray()));
-            System.out.println(Arrays.toString(modulePathEntries.toArray()));
+//            System.out.println(Arrays.toString(annotationRefs.toArray()));
+//            System.out.println(Arrays.toString(modulePathEntries.toArray()));
 
             getMissingEntries(annotationRefs, modulePathEntries)
                     .forEachOrdered(entry -> errorEntry(entry, "Missing annotation for '" + entry + "'"));
@@ -186,7 +216,7 @@ public final class ServiceProviderProcessor extends AbstractProcessor {
     private List<ProviderRef> generateDelegates(List<ProviderRef> refs) {
         for (ProviderRef ref : refs) {
             if (Instantiator.allOf(processingEnv.getTypeUtils(), ref.getService(), ref.getProvider()).stream().anyMatch(o -> o.getKind() == Instantiator.Kind.STATIC_METHOD)) {
-                error(ref, "Static method support not implemented yet");
+                reportError(ref, "Static method support not implemented yet");
             }
         }
         return refs;