+++ java-service-processor/src/test/java/internal/nbbrd/service/provider/ClassPathRegistryTest.java
@@ -17,7 +17,7 @@
 package internal.nbbrd.service.provider;
 
 import static org.assertj.core.api.Assertions.*;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 /**
  *