+++ java-service-processor/src/test/java/internal/nbbrd/service/provider/ModulePathRegistryTest.java
@@ -20,7 +20,7 @@ import internal.nbbrd.service.ModuleInfoEntries;
 import static internal.nbbrd.service.provider.ModulePathRegistry.parseAll;
 import java.io.IOException;
 import static org.assertj.core.api.Assertions.*;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 /**
  *